Swapped floats for doubles

+++ src/main/kotlin/hr/fer/zemris/demo/1.Starfield.kt
@@ -17,7 +17,7 @@ fun main() {
     var prevTime = System.nanoTime()
     while(true) {
         val currTime = System.nanoTime()
-        val delta = (currTime - prevTime).toFloat() / NANO_IN_SEC
+        val delta = (currTime - prevTime).toDouble() / NANO_IN_SEC
         prevTime = currTime
 
         starField.render(canvas, delta)
@@ -35,17 +35,17 @@ class StarField(
     /**
      * Start location of stars should be in [-1, 1] range for x and y and [0, 1] for z
      */
-    private val starsX: FloatArray = FloatArray(starsCount) { randomXStarPosition() }
-    private val starsY: FloatArray = FloatArray(starsCount) { randomYStarPosition() }
-    private val starsZ: FloatArray = FloatArray(starsCount) { randomZStarPosition() }
+    private val starsX: DoubleArray = DoubleArray(starsCount) { randomXStarPosition() }
+    private val starsY: DoubleArray = DoubleArray(starsCount) { randomYStarPosition() }
+    private val starsZ: DoubleArray = DoubleArray(starsCount) { randomZStarPosition() }
 
     private val tanHalfFov: Float = tan(Math.toRadians(fov / 2.0)).toFloat()
 
-    fun render(canvas: Canvas, delta: Float) {
+    fun render(canvas: Canvas, delta: Double) {
         canvas.clear(Color.BLACK)
 
-        val halfWidth = canvas.width / 2.0f
-        val halfHeight =  canvas.height / 2.0f
+        val halfWidth = canvas.width / 2.0
+        val halfHeight =  canvas.height / 2.0
 
         for (i in 0 until starsCount) {
             starsZ[i] -= (delta * speed)
@@ -55,7 +55,7 @@ class StarField(
         }
     }
 
-    private fun drawStar(canvas: Canvas, index: Int, halfWidth: Float, halfHeight: Float) {
+    private fun drawStar(canvas: Canvas, index: Int, halfWidth: Double, halfHeight: Double) {
         val x = (starsX[index]/(starsZ[index] * tanHalfFov) * halfWidth + halfWidth).toInt()
         val y = (starsY[index]/(starsZ[index] * tanHalfFov) * halfHeight + halfHeight).toInt()
 
@@ -72,7 +72,7 @@ class StarField(
         starsZ[index] = randomZStarPosition()
     }
 
-    private fun randomXStarPosition() = (Math.random() - 0.5).toFloat() * 2 * spread
-    private fun randomYStarPosition() = (Math.random() - 0.5).toFloat() * 2 * spread
-    private fun randomZStarPosition() = (Math.random() - 0.00001).toFloat() * spread
+    private fun randomXStarPosition() = (Math.random() - 0.5) * 2 * spread
+    private fun randomYStarPosition() = (Math.random() - 0.5) * 2 * spread
+    private fun randomZStarPosition() = (Math.random() - 0.00001) * spread
 }
