+++ src/main/kotlin/hr/fer/zemris/graphicsAlgorithms/BarycentricCoordinatesCalculator.kt
@@ -11,11 +11,11 @@ object BarycentricCoordinatesCalculator {
      * Equation https://codeplea.com/triangular-interpolation
      */
     fun calculateBarycentricCoordinate(triangle: Triangle, x: Int, y: Int) = with(triangle) {
-        val w1 = ((p2.y - p3.y) * (x - p3.x) + (p3.x - p2.x) * (y - p3.y)).toFloat() /
-                ((p2.y - p3.y) * (p1.x - p3.x) + (p3.x - p2.x) * (p1.y - p3.y)).toFloat()
+        val w1 = ((p2.y - p3.y) * (x - p3.x) + (p3.x - p2.x) * (y - p3.y)).toDouble() /
+                ((p2.y - p3.y) * (p1.x - p3.x) + (p3.x - p2.x) * (p1.y - p3.y)).toDouble()
 
-        val w2 = ((p3.y - p1.y) * (x - p3.x) + (p1.x - p3.x) * (y - p3.y)).toFloat() /
-                ((p2.y - p3.y) * (p1.x - p3.x) + (p3.x - p2.x) * (p1.y - p3.y)).toFloat()
+        val w2 = ((p3.y - p1.y) * (x - p3.x) + (p1.x - p3.x) * (y - p3.y)).toDouble() /
+                ((p2.y - p3.y) * (p1.x - p3.x) + (p3.x - p2.x) * (p1.y - p3.y)).toDouble()
 
         val w3 = 1 - w1 - w2
 
@@ -23,7 +23,7 @@ object BarycentricCoordinatesCalculator {
     }
 
     fun interpolateColorByte(c1: Byte, c2: Byte, c3: Byte, barycentricCoordinates: BarycentricCoordinates): Byte = with(barycentricCoordinates) {
-        (c1.toUnsignedInt().toFloat() * w1 + c2.toUnsignedInt().toFloat() * w2 + c3.toUnsignedInt().toFloat() * w3)
+        (c1.toUnsignedInt().toDouble() * w1 + c2.toUnsignedInt().toDouble() * w2 + c3.toUnsignedInt().toDouble() * w3)
             .roundToInt()
             .toByte()
     }
@@ -31,7 +31,7 @@ object BarycentricCoordinatesCalculator {
 
 
 data class BarycentricCoordinates(
-    val w1: Float,
-    val w2: Float,
-    val w3: Float
+    val w1: Double,
+    val w2: Double,
+    val w3: Double
 )
