+++ src/main/kotlin/hr/fer/zemris/math/extensions/DoubleMathExtensions.kt
@@ -3,8 +3,8 @@ package hr.fer.zemris.math.extensions
 import hr.fer.zemris.math.matrix.Matrix
 import hr.fer.zemris.math.vector.Vector
 
-operator fun Float.times(vector: Vector) =
+operator fun Double.times(vector: Vector) =
     vector * this
 
-operator fun Float.times(matrix: Matrix) =
+operator fun Double.times(matrix: Matrix) =
     matrix * this
